backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Graph processing
networkx==3.2.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
motor==3.3.2
pymongo==4.6.1

# Code analysis
gitpython==3.1.41

# AI/LLM (optional - for explanations)
openai==1.10.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
